+++ ovos_workshop/__init__.py
@@ -2,4 +2,4 @@ from ovos_workshop.app import OVOSAbstractApplication
 from ovos_workshop.decorators import *
 from ovos_workshop.decorators.killable import killable_event, \
     AbortEvent, AbortQuestion
-from ovos_workshop.skills.layers import IntentLayers
+from ovos_workshop.decorators.layers import IntentLayers
