+++ ovos_workshop/decorators/ocp.py
@@ -1,5 +1,5 @@
 from functools import wraps
-from ovos_workshop.skills.layers import IntentLayers
+from ovos_workshop.decorators.layers import IntentLayers
 from ovos_plugin_common_play.ocp import *
 from ovos_plugin_common_play.ocp.status import *
